Comms working with libuarte Nozzle detection OK Ident packet from Console operating
TODO - Modify df_libuarte_drv to work with SOH // EOF packets and multi-byte buffers
TODO - Work out why pulse counting is irregular

+++ df_routines.c
@@ -23,7 +23,7 @@ uint8_t inc_is_error(uint8_t *cnt, uint8_t threshold)
     return(ret);
 }
 
-void wrapping_inc(uint32_t *val, uint32_t lower, uint32_t upper)
+void wrapping_inc(uint8_t *val, uint32_t lower, uint32_t upper)
 {   *val += 1;         
     if (*val >= upper) *val = lower;
 }
